Compute hash collisions for OS X
Each input that produces the same output if converted to a ASCII
vector and multiplied with the matrix
  1   1   1   1   1   1   1   1   1   1   1   1
 11  10   9   8   7   6   5   4   3   2   1   0
 55  45  36  28  21  15  10   6   3   1   0   0
165 120  84  56  35  20  10   4   1   0   0   0
330 210 126  70  35  15   5   1   0   0   0   0
462 252 126  56  21   6   1   0   0   0   0   0
462 210  84  28   7   1   0   0   0   0   0   0
330 120  36   8   1   0   0   0   0   0   0   0
will cause a hash collision on OS X.

This is still just a slow brute-force like search. There would be
a set of equivilent outputs (basically output vectors that map to
the same result) which should be added to compute even more
collisions.

The method is currently limited to printable ASCII characters.